refactor transform record api helpers
